refactor: restructure folders
