+++ src/features/Filter/FilterSlice.ts
@@ -1,6 +1,6 @@
 import { createSlice, PayloadAction } from '@reduxjs/toolkit';
 import { IFilterItem, ITrack } from '../../models';
-import { RootState } from '../../app/store';
+import { RootState } from '../../store';
 
 export interface IFilterSlice {
   field: FieldNames;
