+++ src/features/LoginForm/LoginForm.tsx
@@ -9,10 +9,10 @@ import { Button } from '../Button/Button'
 import './LoginForm.css'
 import { InputField } from '../InputField/InputField'
 import { useNavigate } from 'react-router-dom'
-import { ISignupUser, useUserSignupMutation, useUserTokenMutation } from '../../app/MusicPlayer/music-player.api'
+import { ISignupUser, useUserSignupMutation, useUserTokenMutation } from '../../slices/music-player.api'
 import { useCookies } from 'react-cookie'
-import { useAppDispatch, useLogout } from '../../app/hooks'
-import { setToken } from '../../app/Auth/tokenSlice'
+import { useAppDispatch, useLogout } from '../../hooks'
+import { setToken } from '../../slices/tokenSlice'
 import { getErrorMessage, getPasswordErrorMessage, isPasswordsIdentical, isPasswordValid, isUsernameValid } from './validators'
 import { ILoginFormState } from '../../models'
