+++ src/features/Tracks/Tracks.tsx
@@ -6,8 +6,8 @@ import { Search } from '../../features/Search/Search'
 import { SideMenu } from '../../features/SideMenu/SideMenu'
 import { Sidebar } from '../../features/Sidebar/Sidebar'
 import { Player } from '../Player/Player'
-import { useAppSelector } from '../../app/hooks'
-import { selectActiveTrackId } from '../Track/TrackSlice'
+import { useAppSelector } from '../../hooks'
+import { selectActiveTrackId } from '../../slices/trackSlice'
 import { ITrack } from '../../models'
 
 import { cnTracks } from './Tracks.classname'
