+++ src/index.tsx
@@ -3,7 +3,7 @@ import { Provider } from 'react-redux'
 import { BrowserRouter } from 'react-router-dom'
 import { CookiesProvider } from 'react-cookie'
 
-import { store } from './app/store'
+import { store } from './store'
 import { Routes } from './routes'
 import './index.css'
